backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.58.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
slowapi>=0.1.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
